2024 day 11
These are two completely different solutions. Part 1 is done with linked lists because I thought the ordering was important to maintain. It's reasonable for the first pass.

Part 2 uses a counter and just tracks a count for each number present. This is MUCH faster because we only have to transform a given number once per iteration. Even if we were to increase the number of iterations, it's still fast. 75 iterations is 50ms, 1000 is 1.5 seconds, 2000 is 3 seconds.