2018 Day 22
This also includes an implementation of a priority queue that can have items removed from it (to update priority, for example).

The unbounded map adds 5 seconds of processing time, because it makes the map so wide. But we don't have a good way to know how much larger the map needs to be.