2024 Day 13
I got wrapped around the axle a bit on part 1 trying to solve the system of equations myself. Wolfram Alpha is your friend, once you remember that it likes to solve those for x and y. Once I did that, part 2 was trivial. So only one script for both today.