2024 Day 15
Part 1 was pretty easy to do. Part 2 got a lot more complex in the logic required, but it was still overall straightforward. My first pass I tried to write one method to do both lateral and vertical moves, but it was just too complicated to debug. Separating out lateral moves (which are easy, like part1) and vertical moves (which is where the complexity comes in) made it easier to find the bugs.