2024 Day 16
Part 1 was a fairly simple lowest cost path problem. The only twist was that changing direction was a move, so we needed to track location and facing direction.

Part 2 just required that path tracking be included as well, so that we could backtrack all lowest score paths and total up the number of nodes visited. I just added this to the part 1 code