2019 Day 7
I really expected part 2 to be a super long loop that would break brute force, but nope. The only trick was to add the ability to pause the intcode computer on output so we can add a new input and restart it. To do this, I converted Intcode to a class