2024 Day 21
This took me a while to get to a solution, but I'm very happy with what I got - the part 2 solution runs in under 400us for an overall sequence length of over 485 billion moves. The trick to getting it to run quickly, besides memoization, was to not calculate the actual sequence strings, just the length. Once I stopped working with strings the compute time dropped from days to microseconds.