2024 Day 22
Pretty straightforward. Part 1 was just to make sure that you were doing the math right. Lots of math (a little over 46m iterations), and runs in under 700ms

Part 2 required you to realize that you can't generate everything and then go looking for sequences. That takes too long. Instead, I generate a lookup table of sequences and totals as the numbers are generated (so only once through the loop of buyers * 2000).

It's still not fast at 3.2 seconds, but most of that time is coming from the tuple conversion (350ms) and the dictionary operations (1.4s). Tests with other ways of handling the lookup were all slower.