2019 Day 13
This was a fun one. You could probably write some code to autoplay the game for part 2, but I decided to leave it as a game that I had to play. The only thing I added to that was that it records my keystrokes so that I could resume from a loss.