2024 Day 24
This one took me a long time for part 2. I needed to refactor the way I stored wires and gates from my original part 1 to use classes so that I could cache values. Then I spent a long time trying to figure out why the behavior of the program was not consistent. This was due to the assumption that the x/y value state stayed the same after recursion.

This won't solve for a general case of swapped wires, because it does assume that no more than a single pair of wires needs to be swapped to fix a given bit. This appears to be a valid assumption for our input (and others)